Convert models/MatrixAction to Typescript
