+++ src/bridge/IrcHandler.js
@@ -6,7 +6,7 @@ const { BridgeRequest } = require("../models/BridgeRequest");
 const { IrcRoom } = require("../models/IrcRoom");
 const MatrixRoom = require("matrix-appservice-bridge").MatrixRoom;
 const MatrixUser = require("matrix-appservice-bridge").MatrixUser;
-const MatrixAction = require("../models/MatrixAction");
+const { MatrixAction } = require("../models/MatrixAction");
 const Queue = require("../util/Queue.js");
 const QueuePool = require("../util/QueuePool.js");
 const QuitDebouncer = require("./QuitDebouncer.js");
