+++ src/bridge/MatrixHandler.js
@@ -5,8 +5,8 @@ const Promise = require("bluebird");
 const stats = require("../config/stats");
 const MatrixRoom = require("matrix-appservice-bridge").MatrixRoom;
 const { IrcRoom } = require("../models/IrcRoom");
-const MatrixAction = require("../models/MatrixAction");
 const IrcAction = require("../models/IrcAction");
+const { MatrixAction } = require("../models/MatrixAction");
 const { IrcClientConfig } = require("../models/IrcClientConfig");
 const MatrixUser = require("matrix-appservice-bridge").MatrixUser;
 const { BridgeRequest } = require("../models/BridgeRequest");
@@ -43,6 +43,8 @@ function MatrixHandler(ircBridge, config) {
     this.metrics = {
         //domain => {"metricname" => value}
     };
+    // The media URL to use to transform mxc:// URLs when handling m.room.[file|image]s
+    this._mediaUrl = ircBridge.config.homeserver.media_url || ircBridge.config.homeserver.url;
 }
 
 // ===== Matrix Invite Handling =====
@@ -1229,11 +1231,9 @@ MatrixHandler.prototype._onMessage = Promise.coroutine(function*(req, event) {
         return BridgeRequest.ERR_VIRTUAL_USER;
     }
 
-    // The media URL to use to transform mxc:// URLs when handling m.room.[file|image]s
-    let mediaUrl = this.ircBridge.config.homeserver.media_url;
 
     let mxAction = MatrixAction.fromEvent(
-        this.ircBridge.getAppServiceBridge().getClientFactory().getClientAs(), event, mediaUrl
+        event, this._mediaUrl
     );
     let ircAction = IrcAction.fromMatrixAction(mxAction);
     let ircRooms = yield this.ircBridge.getStore().getIrcChannelsForRoomId(event.room_id);
@@ -1411,9 +1411,6 @@ MatrixHandler.prototype._sendIrcAction = Promise.coroutine(
         req.log.error("Failed to upload text file ", err);
     }
 
-    // The media URL to use to transform mxc:// URLs when handling m.room.[file|image]s
-    let mediaUrl = this.ircBridge.config.homeserver.media_url;
-
     // This is true if the upload was a success
     if (result.content_uri) {
         // Alter event object so that it is treated as if a file has been uploaded
@@ -1422,8 +1419,7 @@ MatrixHandler.prototype._sendIrcAction = Promise.coroutine(
         event.content.body = "sent a long message: ";
 
         // Create a file event to reflect the recent upload
-        let cli = this.ircBridge.getAppServiceBridge().getClientFactory().getClientAs();
-        let mAction = MatrixAction.fromEvent(cli, event, mediaUrl);
+        let mAction = MatrixAction.fromEvent(event, this._mediaUrl);
         let bigFileIrcAction = IrcAction.fromMatrixAction(mAction);
 
         // Replace "Posted a File with..."
@@ -1449,7 +1445,7 @@ MatrixHandler.prototype._sendIrcAction = Promise.coroutine(
             MatrixAction.fromEvent(
                 this.ircBridge.getAppServiceBridge().getClientFactory().getClientAs(),
                 event,
-                mediaUrl
+                this._mediaUrl
             )
         );
